requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dlib==19.18.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
idna==3.4
importlib-metadata==6.6.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.25.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
Pillow==9.5.0
protobuf==3.20.3
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.15.0
